refactor(team): reorganize team member list command
Move the team member listing command from the `listmember` package to `member/list` and introduce a dedicated `member` subcommand group. This provides a clearer hierarchy for team member management
commands.
