+++ internal/cmd/team/member/list/list_test.go
@@ -1,4 +1,4 @@
-package listmember
+package list
 
 import (
 	"bytes"
