+++ internal/cmd/team/team.go
@@ -5,7 +5,7 @@ import (
 	"github.com/tmeckel/azdo-cli/internal/cmd/team/create"
 	"github.com/tmeckel/azdo-cli/internal/cmd/team/delete"
 	"github.com/tmeckel/azdo-cli/internal/cmd/team/list"
-	"github.com/tmeckel/azdo-cli/internal/cmd/team/listmember"
+	"github.com/tmeckel/azdo-cli/internal/cmd/team/member"
 	"github.com/tmeckel/azdo-cli/internal/cmd/team/show"
 	"github.com/tmeckel/azdo-cli/internal/cmd/team/update"
 	"github.com/tmeckel/azdo-cli/internal/cmd/util"
@@ -23,8 +23,8 @@ func NewCmd(ctx util.CmdContext) *cobra.Command {
 
 	cmd.AddCommand(create.NewCmd(ctx))
 	cmd.AddCommand(delete.NewCmd(ctx))
-	cmd.AddCommand(listmember.NewCmd(ctx))
 	cmd.AddCommand(list.NewCmd(ctx))
+	cmd.AddCommand(member.NewCmd(ctx))
 	cmd.AddCommand(show.NewCmd(ctx))
 	cmd.AddCommand(update.NewCmd(ctx))
